chore(Cleanup and restructure the code to be consistent):
