+++ internal/api/v1/service/catalog/index.go
@@ -9,36 +9,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package service
+package catalog
 
 import (
-	"strings"
-
 	"github.com/epinio/epinio/helpers/kubernetes"
 	"github.com/epinio/epinio/internal/api/v1/response"
-	"github.com/epinio/epinio/internal/cli/server/requestctx"
 	"github.com/epinio/epinio/internal/services"
-	"github.com/gin-gonic/gin"
-
 	apierror "github.com/epinio/epinio/pkg/api/core/v1/errors"
-	"github.com/epinio/epinio/pkg/api/core/v1/models"
+	"github.com/gin-gonic/gin"
 )
 
-// Match handles the API endpoint /catalogservicesmatches/:pattern (GET)
-// It returns a list of all Epinio-controlled catalog services matching the prefix pattern.
-func CatalogMatch(c *gin.Context) apierror.APIErrors {
+// Index handles GET /catalogservices
+func Index(c *gin.Context) apierror.APIErrors {
 	ctx := c.Request.Context()
-	log := requestctx.Logger(ctx)
-
-	log.Infow("match catalog services")
-	defer log.Infow("return")
 
 	cluster, err := kubernetes.GetCluster(ctx)
 	if err != nil {
 		return apierror.InternalError(err)
 	}
 
-	log.Infow("list catalog services")
 	kubeServiceClient, err := services.NewKubernetesServiceClient(cluster)
 	if err != nil {
 		return apierror.InternalError(err)
@@ -49,21 +38,6 @@ func CatalogMatch(c *gin.Context) apierror.APIErrors {
 		return apierror.InternalError(err)
 	}
 
-	log.Infow("get service prefix")
-	prefix := c.Param("pattern")
-
-	log.Infow("match prefix", "pattern", prefix)
-	matches := []string{}
-	for _, service := range serviceList {
-		if strings.HasPrefix(service.Meta.Name, prefix) {
-			matches = append(matches, service.Meta.Name)
-		}
-	}
-
-	log.Infow("deliver matches", "found", matches)
-
-	response.OKReturn(c, models.CatalogMatchResponse{
-		Names: matches,
-	})
+	response.OKReturn(c, serviceList)
 	return nil
 }
