+++ internal/api/v1/service/catalog/match.go
@@ -1,0 +1,62 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package catalog
+
+import (
+	"strings"
+
+	"github.com/epinio/epinio/helpers/kubernetes"
+	"github.com/epinio/epinio/internal/api/v1/response"
+	"github.com/epinio/epinio/internal/cli/server/requestctx"
+	"github.com/epinio/epinio/internal/services"
+	apierror "github.com/epinio/epinio/pkg/api/core/v1/errors"
+	"github.com/epinio/epinio/pkg/api/core/v1/models"
+	"github.com/gin-gonic/gin"
+)
+
+// Match handles GET /catalogservicesmatches/:pattern
+func Match(c *gin.Context) apierror.APIErrors {
+	ctx := c.Request.Context()
+	log := requestctx.Logger(ctx)
+
+	log.Infow("match catalog services")
+	defer log.Infow("return")
+
+	cluster, err := kubernetes.GetCluster(ctx)
+	if err != nil {
+		return apierror.InternalError(err)
+	}
+
+	kubeServiceClient, err := services.NewKubernetesServiceClient(cluster)
+	if err != nil {
+		return apierror.InternalError(err)
+	}
+
+	serviceList, err := kubeServiceClient.ListCatalogServices(ctx)
+	if err != nil {
+		return apierror.InternalError(err)
+	}
+
+	prefix := c.Param("pattern")
+
+	log.Infow("match prefix", "pattern", prefix)
+	matches := []string{}
+	for _, svc := range serviceList {
+		if strings.HasPrefix(svc.Meta.Name, prefix) {
+			matches = append(matches, svc.Meta.Name)
+		}
+	}
+
+	log.Infow("deliver matches", "found", matches)
+	response.OKReturn(c, models.CatalogMatchResponse{Names: matches})
+	return nil
+}
