+++ internal/api/v1/service/catalog/show.go
@@ -9,41 +9,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package service
+package catalog
 
 import (
 	"github.com/epinio/epinio/helpers/kubernetes"
 	"github.com/epinio/epinio/internal/api/v1/response"
 	"github.com/epinio/epinio/internal/services"
-	"github.com/gin-gonic/gin"
-
 	apierror "github.com/epinio/epinio/pkg/api/core/v1/errors"
+	"github.com/gin-gonic/gin"
 	k8sapierrors "k8s.io/apimachinery/pkg/api/errors"
 )
 
-func Catalog(c *gin.Context) apierror.APIErrors {
-	ctx := c.Request.Context()
-
-	cluster, err := kubernetes.GetCluster(ctx)
-	if err != nil {
-		return apierror.InternalError(err)
-	}
-
-	kubeServiceClient, err := services.NewKubernetesServiceClient(cluster)
-	if err != nil {
-		return apierror.InternalError(err)
-	}
-
-	serviceList, err := kubeServiceClient.ListCatalogServices(ctx)
-	if err != nil {
-		return apierror.InternalError(err)
-	}
-
-	response.OKReturn(c, serviceList)
-	return nil
-}
-
-func CatalogShow(c *gin.Context) apierror.APIErrors {
+// Show handles GET /catalogservices/:catalogservice
+func Show(c *gin.Context) apierror.APIErrors {
 	ctx := c.Request.Context()
 	serviceName := c.Param("catalogservice")
 
@@ -57,15 +35,14 @@ func CatalogShow(c *gin.Context) apierror.APIErrors {
 		return apierror.InternalError(err)
 	}
 
-	service, err := kubeServiceClient.GetCatalogService(ctx, serviceName)
+	svc, err := kubeServiceClient.GetCatalogService(ctx, serviceName)
 	if err != nil {
 		if k8sapierrors.IsNotFound(err) {
 			return apierror.NewNotFoundError("service instance", serviceName).WithDetails(err.Error())
 		}
-
 		return apierror.InternalError(err)
 	}
 
-	response.OKReturn(c, service)
+	response.OKReturn(c, svc)
 	return nil
 }
